+++ src/main/java/racingcar/domain/car/Cars.java
@@ -1,4 +1,4 @@
-package racingcar.car;
+package racingcar.domain.car;
 
 import static java.util.stream.Collectors.*;
