+++ src/main/java/racingcar/domain/car/MovingPolicy.java
@@ -1,4 +1,4 @@
-package racingcar.car;
+package racingcar.domain.car;
 
 @FunctionalInterface
 public interface MovingPolicy {
